one of the lasts

+++ app/Http/Controllers/ClientsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\User;
 use App\Client;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ClientsController extends Controller
 {
@@ -15,7 +16,11 @@ class ClientsController extends Controller
      */
     public function index()
     {
-        $clients = Client::all();
+		if(Auth::user()->role == 'admin') {
+			$clients = Client::all();
+		} else {
+			$clients = Auth::user()->clients;
+		}
 
         return view('clients.index', ['clients' => $clients]);
     }
@@ -46,11 +51,20 @@ class ClientsController extends Controller
             'activity' => ['required', 'string', 'max:255'],
             'business_address' => ['required', 'string', 'max:255'],
             'home_address' => ['required', 'string', 'max:255'],
-            'maximum_credit' => ['required', 'integer'],
-            'dni' => ['required', 'string', 'unique:clients']
+            'dni' => ['required', 'string', 'unique:clients'],
         ]);
 
-        Client::create($validated);
+		Client::create([
+			'first_name' => $validated['first_name'],
+			'last_name' => $validated['last_name'],
+			'phone' => $validated['phone'],
+			'cell_phone' => $validated['cell_phone'],
+			'activity' => $validated['activity'],
+			'business_address' => $validated['business_address'],
+			'home_address' => $validated['home_address'],
+			'dni' => $validated['dni'],
+			'seller_id' => Auth::user()->id,
+		]);
 
         return redirect('/clients');
     }
@@ -63,6 +77,10 @@ class ClientsController extends Controller
      */
     public function show(Client $client)
     {
+		if(Auth::user()->id != $client->seller_id && Auth::user()->role != 'admin') {
+			return redirect()->back();
+		}
+
         return view('clients.show')->withClient($client);
     }
 
@@ -74,7 +92,13 @@ class ClientsController extends Controller
      */
     public function edit(Client $client)
     {
-        return view('clients.edit', ['client' => $client]);
+		if(Auth::user()->id != $client->seller_id && Auth::user()->role != 'admin') {
+			return redirect()->back();
+		}
+
+		$sellers = User::where('role', '!=', 'disabled')->get();
+
+        return view('clients.edit', ['client' => $client, 'sellers' => $sellers]);
     }
 
     /**
@@ -86,8 +110,11 @@ class ClientsController extends Controller
      */
     public function update(Request $request, Client $client)
     {
-        $validated = $request->validate([
-            'first_name' => ['required', 'string', 'max:255'],
+		if(Auth::user()->id != $client->seller_id && Auth::user()->role != 'admin') {
+			return redirect()->back();
+		}
+
+		$validated = $request->validate([ 'first_name' => ['required', 'string', 'max:255'],
             'last_name' => ['required', 'string', 'max:255'],
             'phone' => ['required', 'string', 'max:255'],
             'cell_phone' => ['required', 'string', 'max:255'],
@@ -95,8 +122,8 @@ class ClientsController extends Controller
             'business_address' => ['required', 'string', 'max:255'],
             'home_address' => ['required', 'string', 'max:255'],
             'maximum_credit' => ['integer'],
-            'max_simultaneous_credits' => ['integer'],
-            'dni' => ['required', 'string']
+            'dni' => ['required', 'string'],
+			'seller_id' => ['required', 'integer'],
         ]);
 
         $checkDni = User::where('dni', $validated['dni'])->get();
@@ -113,9 +140,11 @@ class ClientsController extends Controller
         $client->activity = $validated['activity'];
         $client->business_address = $validated['business_address'];
         $client->home_address = $validated['home_address'];
-        $client->maximum_credit = $validated['maximum_credit'];
-        $client->max_simultaneous_credits = $validated['maximum_credit'];
         $client->dni = $validated['dni'];
+		if(Auth::user()->role == 'admin') {
+			$client->maximum_credit = $validated['maximum_credit'];
+			$client->seller_id = $validated['seller_id'];
+		}
 
         $client->save();
 
@@ -130,6 +159,10 @@ class ClientsController extends Controller
      */
     public function destroy(Client $client)
     {
+        if(Auth::user()->role !== 'admin') {
+            return redirect('/');
+        }
+
         $client->delete();
 
         return redirect('/clients');
@@ -137,28 +170,57 @@ class ClientsController extends Controller
 
     public function search(Request $request)
     {
-        if(isset($request->first_name)) {
-            $validated = $request->validate([
-                'first_name' => ['string', 'max:255']
-            ]);
+		if(Auth::user()->role == 'admin') {
+			if(isset($request->first_name)) {
+				$validated = $request->validate([
+					'first_name' => ['string', 'max:255']
+				]);
 
-            $clients = Client::where('first_name', $validated['first_name'])->get();
-            return view('search.results', ['clients' => $clients]);
-        } elseif(isset($request->last_name)) {
-            $validated = $request->validate([
-                'last_name' => ['string', 'max:255']
-            ]);
+				$clients = Client::where('first_name', $validated['first_name'])->get();
+				return view('search.results', ['clients' => $clients]);
+			} elseif(isset($request->last_name)) {
+				$validated = $request->validate([
+					'last_name' => ['string', 'max:255']
+				]);
 
-            $clients = Client::where('last_name', $validated['last_name'])->get();
-            return view('search.results', ['clients' => $clients]);
-        } elseif(isset($request->dni)) {
-            $validated = $request->validate([
-                'dni' => ['integer']
-            ]);
+				$clients = Client::where('last_name', $validated['last_name'])->get();
+				return view('search.results', ['clients' => $clients]);
+			} elseif(isset($request->dni)) {
+				$validated = $request->validate([
+					'dni' => ['string']
+				]);
 
-            $client = Client::where('dni', $validated['dni'])->first();
-        }
+				$client = Client::where('dni', $validated['dni'])->first();
+			}
+		} else {
+			$clients = Auth::user()->clients;
 
-        return view('credits.create', ['client' => $client]);
+			if(isset($request->first_name)) {
+				$validated = $request->validate([
+					'first_name' => ['string', 'max:255']
+				]);
+
+				$clients = $clients->where('first_name', $validated['first_name']);
+				return view('search.results', ['clients' => $clients]);
+			} elseif(isset($request->last_name)) {
+				$validated = $request->validate([
+					'last_name' => ['string', 'max:255']
+				]);
+
+				$clients = $clients->where('last_name', $validated['last_name']);
+				return view('search.results', ['clients' => $clients]);
+			} elseif(isset($request->dni)) {
+				$validated = $request->validate([
+					'dni' => ['string']
+				]);
+
+				$client = Auth::user()->clients->where('dni', $validated['dni']);
+				if(!isset($client)) {
+					$client = $client[0];
+				}
+			}
+		}
+
+		return view('credits.create', ['client' => $client]);
     }
 }
